# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"RegisterAdminRes": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"AdminName": {
						"type": "string"
					},
					"AdminEmail": {
						"type": "string"
					},
					"AdminPassword": {
						"type": "string"
					},
					"CreatedAt": {
						"type": "string"
					},
					"UpdatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"AdminName",
					"AdminEmail",
					"AdminPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterAdminReq": {
				"properties": {
					"AdminName": {
						"type": "string"
					},
					"AdminEmail": {
						"type": "string"
					},
					"AdminPassword": {
						"type": "string"
					}
				},
				"required": [
					"AdminName",
					"AdminEmail",
					"AdminPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginAdminReq": {
				"properties": {
					"AdminEmail": {
						"type": "string"
					},
					"AdminPassword": {
						"type": "string"
					}
				},
				"required": [
					"AdminEmail",
					"AdminPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"coordinates": {
				"properties": {
					"latitude": {
						"type": "number",
						"format": "double"
					},
					"longitude": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"latitude",
					"longitude"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterResturauntRes": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"ResturauntName": {
						"type": "string"
					},
					"ResturauntContact": {
						"type": "string"
					},
					"location": {
						"$ref": "#/components/schemas/coordinates"
					},
					"CreatedAt": {
						"type": "string"
					},
					"UpdatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"ResturauntName",
					"ResturauntContact",
					"location"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterResturauntReq": {
				"properties": {
					"ResturauntName": {
						"type": "string"
					},
					"ResturauntContact": {
						"type": "string"
					},
					"location": {
						"$ref": "#/components/schemas/coordinates"
					}
				},
				"required": [
					"ResturauntName",
					"ResturauntContact",
					"location"
				],
				"type": "object",
				"additionalProperties": false
			},
			"_LeanDocument__LeanDocument_T__": {
				"properties": {},
				"type": "object"
			},
			"Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/_LeanDocument__LeanDocument_T__",
						"description": "This documents _id."
					},
					"__v": {
						"description": "This documents __v."
					},
					"id": {
						"description": "The string version of this documents _id."
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_": {
				"$ref": "#/components/schemas/Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"LeanDocument_this_": {
				"$ref": "#/components/schemas/Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_"
			},
			"IResturaunt": {
				"$ref": "#/components/schemas/LeanDocument_this_"
			},
			"DeleteResturauntReq": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetResturauntReq": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NearResturauntRes": {
				"properties": {
					"Name": {
						"type": "string"
					},
					"Distance": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"Name",
					"Distance"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetNearResturauntReq": {
				"properties": {
					"location": {
						"$ref": "#/components/schemas/coordinates"
					},
					"radius": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"location",
					"radius"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterUserRes": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"UserName": {
						"type": "string"
					},
					"UserEmail": {
						"type": "string"
					},
					"UserPassword": {
						"type": "string"
					},
					"UserContact": {
						"type": "string"
					},
					"CreatedAt": {
						"type": "string"
					},
					"UpdatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"UserName",
					"UserEmail",
					"UserPassword",
					"UserContact"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterUserReq": {
				"properties": {
					"UserName": {
						"type": "string"
					},
					"UserEmail": {
						"type": "string"
					},
					"UserPassword": {
						"type": "string"
					},
					"UserContact": {
						"type": "string"
					}
				},
				"required": [
					"UserName",
					"UserEmail",
					"UserPassword",
					"UserContact"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUser": {
				"$ref": "#/components/schemas/LeanDocument_this_"
			},
			"DeleteUserReq": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "resturaunt_info_system",
		"version": "1.0.0",
		"description": "Resturaunt info System",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Shahid Awan"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/Admin/RegisterAdmin": {
			"post": {
				"operationId": "RegisterAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterAdminRes"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterAdminReq"
							}
						}
					}
				}
			}
		},
		"/Admin/LoginAdmin": {
			"post": {
				"operationId": "LoginAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterAdminRes"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginAdminReq"
							}
						}
					}
				}
			}
		},
		"/Resturaunt/RegisterResturaunt": {
			"post": {
				"operationId": "RegisterResturaunt",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterResturauntRes"
								}
							}
						}
					}
				},
				"tags": [
					"Resturaunt"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterResturauntReq"
							}
						}
					}
				}
			}
		},
		"/Resturaunt/DeleteResturaunt": {
			"delete": {
				"operationId": "DeleteResturaunt",
				"responses": {
					"200": {
						"description": "Resturaunt Deleted!",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IResturaunt"
										},
										{
											"properties": {
												"_id": {
													"type": "string"
												}
											},
											"required": [
												"_id"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Resturaunt"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteResturauntReq"
							}
						}
					}
				}
			}
		},
		"/Resturaunt/GetResturaunt": {
			"post": {
				"operationId": "GetResturaunt",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterResturauntRes"
								}
							}
						}
					}
				},
				"tags": [
					"Resturaunt"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetResturauntReq"
							}
						}
					}
				}
			}
		},
		"/Resturaunt/GetResturauntLocation": {
			"post": {
				"operationId": "GetResturauntLocation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/NearResturauntRes"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Resturaunt"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetNearResturauntReq"
							}
						}
					}
				}
			}
		},
		"/User/RegisterUser": {
			"post": {
				"operationId": "RegisterUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterUserRes"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterUserReq"
							}
						}
					}
				}
			}
		},
		"/User/DeleteUser": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "User Deleted!",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IUser"
										},
										{
											"properties": {
												"_id": {
													"type": "string"
												}
											},
											"required": [
												"_id"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteUserReq"
							}
						}
					}
				}
			}
		},
		"/User/GetUserList": {
			"post": {
				"operationId": "GetUserList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RegisterUserRes"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}